\name{plot.hcas}
\alias{plot.hcas}
\title{Plot function for optHCAS output.}
\description{
This function is used to plot optimal sample coordinates.
}
\usage{
\method{plot}{hcas}(x, ...)
}
\arguments{
\item{x}{Object of class inheriting from "hcas".}
\item{...}{Other arguments.}
}
\value{
}
\seealso{
\code{\link{runOpt}}.
}
\examples{
\dontrun{
##

plot(out) # where out is the output from hcas class

##
}
}
\keyword{hcas}
